generalize backup sequence
